config should use avalue_t directly, ditch struct avalue
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@18005 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ willow/src/include/config.h
@@ -80,8 +80,8 @@ extern struct configuration : noncopyable {
 	bool		 x_follow;
 	int		 max_redirects;
 	bool		 use_dio;
-	long		 cache_memory;
-	long		 max_entity_size;
+	size_t		 cache_memory;
+	size_t		 max_entity_size;
 	string		 cache_master;
 	bool		 htcp_sigrequired;
